+++ projects/ausmittlung-lib/src/lib/services/second-factor-transaction.service.ts
@@ -6,16 +6,13 @@
 
 import { DialogService } from '@abraxas/voting-lib';
 import { Injectable, inject } from '@angular/core';
-import { defer, Observable, of, OperatorFunction, throwError } from 'rxjs';
-import { concatMap, retryWhen } from 'rxjs/operators';
+import { Observable } from 'rxjs';
 import {
   SecondFactorTransactionDialogComponent,
   SecondFactorTransactionDialogData,
+  SecondFactorTransactionDialogResult,
 } from '../components/transaction-request-dialog/second-factor-transaction-dialog.component';
-import { isErrorType } from './utils/error.utils';
-
-const ERROR_TYPE_VERIFY_SECOND_FACTOR = 'VerifySecondFactorTimeoutException';
-const RETRY_COUNT = 5;
+import { SecondFactorTransactionNevisInfo, SecondFactorTransactionProvider } from '../models';
 
 @Injectable({
   providedIn: 'root',
@@ -23,52 +20,28 @@ const RETRY_COUNT = 5;
 export class SecondFactorTransactionService {
   private readonly dialog = inject(DialogService);
 
-  public showDialogAndExecuteVerifyAction<T>(action: () => Observable<T>, code: string, qrCode: string): Promise<void> {
+  public async showDialogAndExecuteVerifyAction<T>(
+    action: (otpCode?: string) => Observable<T>,
+    nevisInfo?: SecondFactorTransactionNevisInfo,
+    availableProviders?: SecondFactorTransactionProvider[],
+  ): Promise<void> {
     const data: SecondFactorTransactionDialogData = {
-      code: code,
-      qrCode: qrCode,
+      nevisInfo,
+      availableProviders,
+      action,
     };
-    const dialogRef = this.dialog.open<SecondFactorTransactionDialogComponent>(SecondFactorTransactionDialogComponent, data);
 
-    return new Promise<void>((resolve, reject) => {
-      const subscription = action()
-        .pipe(this.retryOnVerifyTimeout())
-        .subscribe(
-          () => {
-            dialogRef.close();
-            resolve();
-          },
-          err => {
-            dialogRef.componentInstance.hasError = true;
-            reject(err);
-          },
-        );
+    const result = await this.dialog.openForResult<SecondFactorTransactionDialogComponent, SecondFactorTransactionDialogResult>(
+      SecondFactorTransactionDialogComponent,
+      data,
+    );
 
-      dialogRef.afterClosed().subscribe(() => {
-        subscription.unsubscribe();
-        reject();
-      });
-    });
-  }
+    if (result?.error) {
+      throw result.error;
+    }
 
-  private retryOnVerifyTimeout<T>(): OperatorFunction<T, T> {
-    return (src: Observable<T>) =>
-      defer(() => {
-        // retry on failure after timeout
-        return src.pipe(
-          retryWhen(errs =>
-            errs.pipe(
-              concatMap((err, index) => {
-                // after timeout retry x times (see RETRY_COUNT)
-                if (isErrorType(err, ERROR_TYPE_VERIFY_SECOND_FACTOR) && index < RETRY_COUNT) {
-                  return of(null);
-                }
-
-                return throwError(err);
-              }),
-            ),
-          ),
-        );
-      });
+    if (!result?.verified) {
+      throw new Error('Second factor transaction not verified');
+    }
   }
 }